refactor: reorganize training models into modular base class structure with specialized implementations
