+++ src/tasks/training/fabric/model/utils.py
@@ -1,0 +1,7 @@
+from transformers import AutoModelForCausalLM, AutoModelForMaskedLM
+
+
+AVAILABLE_MODELS = (
+    AutoModelForCausalLM,
+    AutoModelForMaskedLM
+)
